Refactor pathfinding and utils, add tests
