+++ .eslintrc
@@ -19,7 +19,7 @@
 		]
 	},
 	"overrides": [{
-		"files": "./test/*.js",
+		"files": "./test/*/*.js",
 		"rules": {
 			"node/no-unpublished-require": 0,
 			"node/no-missing-require": 0
