+++ routes/bots.js
@@ -4,7 +4,7 @@ const botsRouter = express.Router();
 
 let { BruinBot, InventoryArticle } = require('../models/bruinbot.model');
 let { Path } = require('../models/map.model');
-let util = require('./utils');
+let util = require('../util/utils');
 
 /**
  * ----------------- GET (return information about objects) ----------------
