+++ test/routes/assets/sample-image.jpg
